+++ R/dd_calc.R
@@ -1,4 +1,4 @@
-#' Compute degree days
+#' Compute degree days from daily data
 #'
 #' Compute degree days from daily min and max temperature
 #'
@@ -10,7 +10,9 @@
 #' @param method The method used
 #' @param cutoff The cutoff method
 #' @param cumulative Return cumulative values
-#' @param noneg Set negative values to zero
+#' @param no_neg Set negative values to zero
+#' @param interpolate_na Interpolate missing values
+#' @param quiet Suppress messages
 #' @param debug Show additional messages
 #'
 #' @details Units for \code{daily_min}, \code{daily_max}, \code{thresh_low}, and \code{thresh_up} should all be the same
@@ -19,19 +21,25 @@
 #' \code{nextday_min} is required for the double-triangle and the double-sine methods. These methods use the minimum temperature
 #' of the following day to model temperatures in the 2nd half of thd day.
 #'
-#' \code{noneg = TRUE} sets negative values to zero. This is generally accepted when using degree days to predict the timing of
+#' \code{no_neg = TRUE} sets negative values to zero. This is generally preferred when using degree days to predict the timing of
 #'  development milestones, if one assumes that growth can not go backwards.
 #'
+#' Missing values (NAs) in the temperatures will result in NA degree days. If \code{interpolate_na = TRUE}, missing degree days will
+#' be interpolated. NAs in the middle of the series will be linearly interpolated, and NAs at the ends will be filled with the
+#' adjacent values.
+#'
 #' @importFrom crayon yellow red
+#' @importFrom zoo na.fill
+#' @importFrom methods is
 #' @export
 
-dd_vals <- function(daily_min, daily_max, nextday_min = NULL,
+dd_calc <- function(daily_min, daily_max, nextday_min = NULL,
                      thresh_low = NULL, thresh_up = NULL,
                 method = c("sng_tri", "dbl_tri", "sng_sine", "dbl_sine", "simp_avg")[0],
                 cutoff = c("horizontal", "vertical")[1],
                 cumulative = FALSE,
-                noneg = TRUE,
-                debug = FALSE) {
+                no_neg = TRUE, interpolate_na = FALSE,
+                quiet = FALSE, debug = FALSE) {
 
   if (!method %in% c("sng_tri", "dbl_tri", "sng_sine", "dbl_sine", "simp_avg")) stop("unknown value for `method`")
   if (!cutoff %in% c("horizontal", "vertical")) stop("unknown value for `cutoff`")
@@ -41,33 +49,47 @@ dd_vals <- function(daily_min, daily_max, nextday_min = NULL,
 
   if (FALSE %in% (daily_min <= daily_max)) stop("daily_min must always be less than daily_max")
 
-  ## Set default values for lower and upper threshold, so we can define the case.
-  ## These will be further checked below
-  # thresh_up_use <- ifelse(is.null(thresh_up), 200, thresh_up)
-  # thresh_low_use <- ifelse(is.null(thresh_low), 0, thresh_low)
+  ## Get rid of the units. (alternatively I would need to check that thresh_low and thresh_up have the
+  ## same units, otherwise some of the equality tests below will fail)
+  if (is(daily_min, "units")) daily_min <- as.numeric(daily_min)
+  if (is(daily_max, "units")) daily_max <- as.numeric(daily_max)
 
-  ## Initialize an object to hold the results
+  ## Create a template error message
+  err_missing_thresh <- "The {method} method requires you pass lower and upper threshhold. (If the parameters you were given do not include an upper threshold, pass a large number like 200.)"
+
+  ## To handle NAs, find the indices where we have valid temps
+  good_idx <- which(!is.na(daily_min) & !is.na(daily_max))
+
+  if (length(good_idx) == 0) {
+    stop("Invalid values for daily_min and/or daily_max (all NAs)")
+  } else if (length(daily_min) != length(good_idx)) {
+    if (!quiet) message(yellow(" - NAs found in daily_min and/or daily_max."))
+    if (cumulative && !interpolate_na) stop(" - Can't return cumlative degree days because there are NAs in the input temperatures. You can let `interpolate_na = TRUE`, or fill the missing temperature values and try again.")
+  }
+
+  ## Initialize objects to hold results
+  dd_all <- rep(NA, length(daily_min))  ## NAs will be replaced with computed dd at the very end
   degday <- NULL
 
   if (method == "simp_avg") {
     ### ------ SIMPLE AVERAGE -------
     #################################
 
-    if (debug) message(yellow(" - using simple average method"))
-    if (!is.null(thresh_up)) message(yellow(" - the simple average method doesn't support an upper threshhold, ignoring"))
+    if (!quiet) message(yellow(" - using simple average method"))
+    if (!is.null(thresh_up) && !quiet) message(yellow(" - the simple average method doesn't support an upper threshhold, ignoring"))
     thresh_low_use <- ifelse(is.null(thresh_low), 0, thresh_low)
-    degday <- ((daily_max + daily_min) / 2) - thresh_low_use
+    degday <- ((daily_max[good_idx] + daily_min[good_idx]) / 2) - thresh_low_use
 
   } else if (method == "sng_tri") {
     ### ------ SINGLE TRIANGLE -------
     ### Formulas come from Table 3 in Zalom et al 1983
 
-    if (debug) message(yellow(" - using single triangle method"))
+    if (!quiet) message(yellow(" - using single triangle method"))
     if (is.null(thresh_low) || is.null(thresh_up)) {
-      stop(red("The single triangle method requires you pass lower and upper threshhold"))
+      stop(red(sub("\\{method\\}", "single triangle", err_missing_thresh)))
     }
 
-    for (i in 1:length(daily_min)) {
+    for (i in good_idx) {
       this_case <- dd_case(daily_min[i], daily_max[i], thresh_low, thresh_up)
       if (debug) message(yellow(paste0(" - ", i, ": case ", this_case)))
       degday <- c(degday,
@@ -86,10 +108,10 @@ dd_vals <- function(daily_min, daily_max, nextday_min = NULL,
     ### ------ DOUBLE TRIANGLE (HALF DAY FORMULAS) -------
     ### Formulas come from Table 4 in Zalom et al 1983
 
-    if (debug) message(yellow(" - using double triangle method"))
+    if (!quiet) message(yellow(" - using double triangle method"))
 
     if (is.null(thresh_low) || is.null(thresh_up)) {
-      stop(red("The double triangle method requires you pass lower and upper threshhold"))
+      stop(red(sub("\\{method\\}", "double triangle", err_missing_thresh)))
     }
 
     if (is.null(nextday_min)) {
@@ -98,7 +120,7 @@ dd_vals <- function(daily_min, daily_max, nextday_min = NULL,
 
     if (length(nextday_min) != length(daily_min)) stop("The next day minimum temperature must be the same length as the daily min and max temps")
 
-    for (i in 1:length(daily_min)) {
+    for (i in good_idx) {
 
       ## With the double-triangle half-day formulas, we need to
       ## compute two GDD values for each half day. This requires that
@@ -121,13 +143,13 @@ dd_vals <- function(daily_min, daily_max, nextday_min = NULL,
     ### ------ SINGLE SINE -------
     ### Formulas come from Table 5 in Zalom et al 1983
 
-    if (debug) message(yellow(" - using single sine method"))
+    if (!quiet) message(yellow(" - using single sine method"))
 
     if (is.null(thresh_low) || is.null(thresh_up)) {
-      stop(red("The single sine method requires you pass lower and upper threshhold"))
+      stop(red(sub("\\{method\\}", "single sine", err_missing_thresh)))
     }
 
-    for (i in 1:length(daily_min)) {
+    for (i in good_idx) {
 
       this_case <- dd_case(daily_min[i], daily_max[i], thresh_low, thresh_up)
       if (debug) message(yellow(paste0(" - ", i, ": case = ", this_case)))
@@ -167,10 +189,10 @@ dd_vals <- function(daily_min, daily_max, nextday_min = NULL,
     ### ------ DOUBLE SINE -------
     ### Formulas come from Table 6 in Zalom et al 1983
 
-    if (debug) message(yellow(" - using double sine method"))
+    if (!quiet) message(yellow(" - using double sine method"))
 
     if (is.null(thresh_low) || is.null(thresh_up)) {
-      stop(red("The double sine method requires you pass lower and upper threshhold"))
+      stop(red(sub("\\{method\\}", "double sine", err_missing_thresh)))
     }
 
     if (is.null(nextday_min)) {
@@ -179,7 +201,7 @@ dd_vals <- function(daily_min, daily_max, nextday_min = NULL,
 
     if (length(nextday_min) != length(daily_min)) stop("The next day minimum temperature must be the same length as the daily min and max temps")
 
-    for (i in 1:length(daily_min)) {
+    for (i in good_idx) {
 
       ## With the double-sine half-day formulas, we need to compute two GDD values for each half day.
 
@@ -199,45 +221,60 @@ dd_vals <- function(daily_min, daily_max, nextday_min = NULL,
 
   }
 
+  ## Zero out negative values
+  if (no_neg && sum(degday < 0) > 0) {
+    if (!quiet) message(yellow(paste0(" - zeroing out ", sum(degday < 0), " negative degree day values")))
+    degday[degday < 0] <- 0
+  }
+
+  ## Swap in the computed values
+  dd_all[good_idx] <- degday
+
+  ## Interpolate missing values
+  if (anyNA(dd_all) && interpolate_na) {
+    if (!quiet) message(yellow(paste0(" - interpolating ", sum(is.na(dd_all)), " missing degree day values")))
+    dd_all <- na.fill(dd_all, "extend")
+  }
+
   ## Return degree days
   if (cumulative) {
-    cumsum(degday)
+    cumsum(dd_all)
   } else {
-    degday
+    dd_all
   }
 
 }
 
-#' @describeIn dd_vals Compute degree days using the single-triangle method
+#' @describeIn dd_calc Compute degree days using the single-triangle method
 #' @export
 
 dd_sng_tri <- function(daily_min, daily_max, thresh_low = NULL, thresh_up = NULL, cumulative = FALSE) {
-  dd_vals(daily_min=daily_min, daily_max=daily_max, thresh_low=thresh_low, thresh_up=thresh_up,
+  dd_calc(daily_min=daily_min, daily_max=daily_max, thresh_low=thresh_low, thresh_up=thresh_up,
           method ="sng_tri", cutoff = "horizontal", cumulative=cumulative)
 }
 
-#' @describeIn dd_vals Compute degree days using the single-sine method
+#' @describeIn dd_calc Compute degree days using the single-sine method
 #' @export
 
 dd_sng_sine <- function(daily_min, daily_max, thresh_low = NULL, thresh_up = NULL, cumulative = FALSE) {
-  dd_vals(daily_min=daily_min, daily_max=daily_max, thresh_low=thresh_low, thresh_up=thresh_up,
+  dd_calc(daily_min=daily_min, daily_max=daily_max, thresh_low=thresh_low, thresh_up=thresh_up,
           method ="sng_sine", cutoff = "horizontal", cumulative=cumulative)
 }
 
-#' @describeIn dd_vals Compute degree days using the double-triangle method
+#' @describeIn dd_calc Compute degree days using the double-triangle method
 #' @export
 
 dd_dbl_tri <- function(daily_min, daily_max, nextday_min = NULL, thresh_low = NULL, thresh_up = NULL, cumulative = FALSE) {
-  dd_vals(daily_min=daily_min, daily_max=daily_max, nextday_min=nextday_min,
+  dd_calc(daily_min=daily_min, daily_max=daily_max, nextday_min=nextday_min,
           thresh_low=thresh_low, thresh_up=thresh_up,
           method ="dbl_tri", cutoff = "horizontal", cumulative=cumulative)
 }
 
-#' @describeIn dd_vals Compute degree days using the double-sine method
+#' @describeIn dd_calc Compute degree days using the double-sine method
 #' @export
 
 dd_dbl_sine <- function(daily_min, daily_max, nextday_min = NULL, thresh_low = NULL, thresh_up = NULL, cumulative = FALSE) {
-  dd_vals(daily_min=daily_min, daily_max=daily_max, nextday_min=nextday_min,
+  dd_calc(daily_min=daily_min, daily_max=daily_max, nextday_min=nextday_min,
           thresh_low=thresh_low, thresh_up=thresh_up,
           method ="dbl_sine", cutoff = "horizontal", cumulative=cumulative)
 }
@@ -284,7 +321,7 @@ dd_case <- function(daily_min, daily_max, thresh_low, thresh_up) {
 }
 
 #' Compute half-day GDD using triangle functions.
-#' This is an internal function, called by dd_vals()
+#' This is an internal function, called by dd_calc()
 #' @keywords internal
 
 dd_dbltri_half <- function(case, tmin, tmax, thresh_low, thresh_up) {
@@ -300,7 +337,7 @@ dd_dbltri_half <- function(case, tmin, tmax, thresh_low, thresh_up) {
 }
 
 #' Compute half-day GDD using sine functions.
-#' This is an internal function, called by dd_vals()
+#' This is an internal function, called by dd_calc()
 #' @keywords internal
 
 dd_dblsine_half <- function(case, tmin, tmax, thresh_low, thresh_up) {