+++ NEWS.md
@@ -1,3 +1,8 @@
+# degday 0.2.0 (2021-12-14)
+
+* added `dd_fetch_cimis()`: fetches daily temperature data from weather stations in the CIMIS network (may move this to another package)
+* `dd_calc()`: added `quiet` and `interpolate_na` arguments; updated to handle NA values
+
 # degday 0.1.0
 
 * Initial commit